Make object layout default to old style and add option to use new grid layout.
Improve look of switcher for multiple and enum editors.
Fix spacing inconsistencies in themes.
Strings and Select editors now determine the optimum grid size using the length of the title/options.

Grid layout is no longer the default and can be enabled with options:
```json
{
  "type": "object",
  "format": "grid"
}
```
or
```json
{
  "type": "object",
  "options": {
    "layout": "grid"
  }
}
```
or
```js
new JSONEditor(el,{
  schema: {type: "object"},
  object_layout: "grid"
});
```